refactor(libs): address review - declare phantom deps, trim tsconfigs, drop redundant configs
Resolves @gudnuf's phantom-dependency review + @jbojcic1's cleanups on #1159.

Phantom deps (now declared; were resolving only via hoisting):
- @agicash/cashu: + @noble/hashes (runtime, blind-signature-matching) as dep,
  + @noble/curves (used in its test) as devDep.
- apps/web-wallet: + @scure/base (base64url in lightning-address-service) and
  + @supabase/realtime-js (type import in the realtime channel builder).
- @scure/base hoisted into the root catalog (now used by bolt11 + web).
- Skipped the flagged `supabase/database.types` - it's a tsconfig path alias to the
  repo-root generated file (import type), not an npm dep.

Cleanups (@jbojcic1):
- Moved the Money DevTools formatter out of @agicash/money into the web app
  (app/lib/money-devtools-formatter.ts) so the money package needs no DOM lib.
- Trimmed each package tsconfig lib/types to what it uses: utils neither; ecies/
  bolt11/cashu drop DOM (keep bun for tests); lnurl keeps DOM (URL) drops bun;
  money drops DOM.
- @agicash/ecies: made the implementation file the entrypoint (ecies.ts -> index.ts),
  dropping the pointless re-export barrel.
- Removed the redundant per-package biome.json (7 files); the root biome.jsonc
  already covers the whole tree (verified).
- Moved type-fest to a devDep on @agicash/cashu (type-only import). typecheck + build
  stay green - it is erased at build and resolves for tsc within the workspace.

Verification: typecheck 9/9, unit tests (money 14, bolt11 9, ecies 18, cashu 35,
web 57), web build, biome - all green.
