+++ apps/web-wallet/package.json
@@ -45,10 +45,12 @@
     "@react-router/express": "7.14.2",
     "@react-router/fs-routes": "7.14.2",
     "@react-router/node": "7.14.2",
+    "@scure/base": "catalog:",
     "@scure/bip32": "1.7.0",
     "@scure/bip39": "1.6.0",
     "@sentry/profiling-node": "10.42.0",
     "@sentry/react-router": "10.42.0",
+    "@supabase/realtime-js": "2.95.2",
     "@supabase/supabase-js": "2.95.2",
     "@tanstack/react-query": "5.90.20",
     "@tanstack/react-query-devtools": "5.91.3",
