+++ package.json
@@ -10,6 +10,7 @@
       "@noble/ciphers": "1.3.0",
       "@noble/curves": "1.9.7",
       "@noble/hashes": "1.8.0",
+      "@scure/base": "2.0.0",
       "@stablelib/base64": "2.0.1",
       "@stablelib/chacha20poly1305": "2.0.1",
       "@types/big.js": "6.2.2",
