+++ packages/lnurl/tsconfig.json
@@ -4,7 +4,7 @@
   "exclude": ["node_modules"],
   "compilerOptions": {
     "lib": ["ES2022", "DOM"],
-    "types": ["bun"],
+    "types": [],
     "noEmit": true
   }
 }
